RatingMapper broken down
